backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.12.1",
    "fastapi==0.104.1",
    "fastapi-cors==0.0.6",
    "numpy==1.26.4",
    "pandas>=2.3.3",
    "pydantic==2.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dotenv==1.0.0",
    "scikit-surprise>=1.1.4",
    "sqlalchemy==2.0.23",
    "uvicorn==0.24.0",
]
